move projectiles config to projectile spaceobject module
