+++ modules/core/src/configurations/index.ts
@@ -1,7 +1,6 @@
 import { dragonflySF22 } from './dragonfly-sf-22';
 
 export * from './dragonfly-sf-22';
-export * from './projectiles';
 
 export const shipConfigurations = {
     'dragonfly-SF22': dragonflySF22,
